Support a JSON deriver for ezjsonm
The existing code for the Yaml deriver could quite easily be reused to
build a JSON deriver using Ezjsonm (they share the same `value` type).
This PR refactors the whole ppx to become a generic library and the
instantiates a Yaml and a Ezjsonm instance of the library.

Additionally the README which contained many tests now resides in odoc
documentation pages.